kata_agent: Update host and guest paths for Kata
Both paths related to the agent should not be into /tmp, and to be
consistent with Clear Containers agent, we should use /run path.

Also, it renames the guest path from ../pods/.. to ../containers/..
since it does not make sense to have "pods" inside the agent which
represents only one pod.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>